project/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
defusedxml==0.7.1
Django==5.1.2
django-admin-interface==0.29.1
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
django-rest-swagger==2.2.0
django-richtextfield==1.6.2
django-simple-history==3.7.0
django-templated-mail==1.1.1
django-ulid==0.0.4
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
simplejson==3.19.3
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
ua-parser==0.18.0
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
uvicorn==0.32.0
